requirements.txt
pylast
python-dotenv